+++ sshd-core/src/main/java/org/apache/sshd/agent/SshAgent.java
@@ -21,7 +21,6 @@ package org.apache.sshd.agent;
 import java.io.IOException;
 import java.security.KeyPair;
 import java.security.PublicKey;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -31,7 +30,7 @@ public interface SshAgent extends java.nio.channels.Channel {
 
     String SSH_AUTHSOCKET_ENV_NAME = "SSH_AUTH_SOCK";
 
-    List<? extends Map.Entry<PublicKey, String>> getIdentities() throws IOException;
+    Iterable<? extends Map.Entry<PublicKey, String>> getIdentities() throws IOException;
 
     byte[] sign(PublicKey key, byte[] data) throws IOException;